Move CodeQL setup to its own file
