+++ lib/codeql.test.js
@@ -24,7 +24,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.stubToolRunnerConstructor = void 0;
 const fs = __importStar(require("fs"));
-const path = __importStar(require("path"));
+const path_1 = __importDefault(require("path"));
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
 const toolcache = __importStar(require("@actions/tool-cache"));
 const safeWhich = __importStar(require("@chrisgavin/safe-which"));
@@ -34,9 +34,8 @@ const yaml = __importStar(require("js-yaml"));
 const nock_1 = __importDefault(require("nock"));
 const sinon = __importStar(require("sinon"));
 const actionsUtil = __importStar(require("./actions-util"));
-const api = __importStar(require("./api-client"));
 const codeql = __importStar(require("./codeql"));
-const defaults = __importStar(require("./defaults.json"));
+const defaults = __importStar(require("./defaults.json")); // Referenced from codeql-action-sync-tool!
 const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
@@ -101,7 +100,7 @@ async function mockDownloadApi({ apiDetails = sampleApiDetails, isPinned, tagNam
         : `/download/${tagName}/codeql-bundle.tar.gz`;
     (0, nock_1.default)(baseUrl)
         .get(relativeUrl)
-        .replyWithFile(200, path.join(__dirname, `/../src/testdata/codeql-bundle${isPinned ? "-pinned" : ""}.tar.gz`));
+        .replyWithFile(200, path_1.default.join(__dirname, `/../src/testdata/codeql-bundle${isPinned ? "-pinned" : ""}.tar.gz`));
     return `${baseUrl}${relativeUrl}`;
 }
 async function installIntoToolcache({ apiDetails = sampleApiDetails, isPinned, tagName, tmpDir, }) {
@@ -238,7 +237,7 @@ for (const [isFeatureEnabled, toolsInput, shouldToolcacheBeBypassed,] of TOOLCAC
         });
         (0, nock_1.default)("https://example.githubenterprise.com")
             .get(`/github/codeql-action/releases/download/${defaults.bundleVersion}/${codeQLBundleName}`)
-            .replyWithFile(200, path.join(__dirname, `/../src/testdata/codeql-bundle-pinned.tar.gz`));
+            .replyWithFile(200, path_1.default.join(__dirname, `/../src/testdata/codeql-bundle-pinned.tar.gz`));
         // This is a workaround to mock `api.getApiDetails()` since it doesn't seem to be possible to
         // mock this directly. The difficulty is that `getApiDetails()` is called locally in
         // `api-client.ts`, but `sinon.stub(api, "getApiDetails")` only affects calls to
@@ -261,28 +260,6 @@ for (const [isFeatureEnabled, toolsInput, shouldToolcacheBeBypassed,] of TOOLCAC
         t.is(cachedVersions.length, 1);
     });
 });
-(0, ava_1.default)("parse codeql bundle url version", (t) => {
-    t.deepEqual(codeql.getCodeQLURLVersion("https://github.com/.../codeql-bundle-20200601/..."), "20200601");
-});
-(0, ava_1.default)("convert to semver", (t) => {
-    const tests = {
-        "20200601": "0.0.0-20200601",
-        "20200601.0": "0.0.0-20200601.0",
-        "20200601.0.0": "20200601.0.0",
-        "1.2.3": "1.2.3",
-        "1.2.3-alpha": "1.2.3-alpha",
-        "1.2.3-beta.1": "1.2.3-beta.1",
-    };
-    for (const [version, expectedVersion] of Object.entries(tests)) {
-        try {
-            const parsedVersion = codeql.convertToSemVer(version, (0, logging_1.getRunnerLogger)(true));
-            t.deepEqual(parsedVersion, expectedVersion);
-        }
-        catch (e) {
-            t.fail(e instanceof Error ? e.message : String(e));
-        }
-    }
-});
 (0, ava_1.default)("getExtraOptions works for explicit paths", (t) => {
     t.deepEqual(codeql.getExtraOptions({}, ["foo"], []), []);
     t.deepEqual(codeql.getExtraOptions({ foo: [42] }, ["foo"], []), ["42"]);
@@ -304,20 +281,6 @@ for (const [isFeatureEnabled, toolsInput, shouldToolcacheBeBypassed,] of TOOLCAC
     t.throws(() => codeql.getExtraOptions({ "*": 42 }, ["foo"], []));
     t.throws(() => codeql.getExtraOptions({ foo: 87 }, ["foo"], []));
     t.throws(() => codeql.getExtraOptions({ "*": [42], foo: { "*": 87, bar: [99] } }, ["foo", "bar"], []));
-});
-(0, ava_1.default)("getCodeQLActionRepository", (t) => {
-    const logger = (0, logging_1.getRunnerLogger)(true);
-    (0, util_1.initializeEnvironment)("1.2.3");
-    // isRunningLocalAction() === true
-    delete process.env["GITHUB_ACTION_REPOSITORY"];
-    process.env["RUNNER_TEMP"] = path.dirname(__dirname);
-    const repoLocalRunner = codeql.getCodeQLActionRepository(logger);
-    t.deepEqual(repoLocalRunner, "github/codeql-action");
-    // isRunningLocalAction() === false
-    sinon.stub(actionsUtil, "isRunningLocalAction").returns(false);
-    process.env["GITHUB_ACTION_REPOSITORY"] = "xxx/yyy";
-    const repoEnv = codeql.getCodeQLActionRepository(logger);
-    t.deepEqual(repoEnv, "xxx/yyy");
 });
 (0, ava_1.default)("databaseInterpretResults() does not set --sarif-add-query-help for 2.7.0", async (t) => {
     const runnerConstructorStub = stubToolRunnerConstructor();
@@ -602,48 +565,6 @@ const injectedConfigMacro = ava_1.default.macro({
     sinon.stub(safeWhich, "safeWhich").resolves("");
     await codeqlObject.databaseInterpretResults("", [], "", "", "", "-v", "");
     t.false(runnerConstructorStub.firstCall.args[1].includes("--sarif-add-baseline-file-info"), "--sarif-add-baseline-file-info must be absent, but it is present");
-});
-(0, ava_1.default)("findCodeQLBundleTagDotcomOnly() matches GitHub Release with marker file", async (t) => {
-    // Look for GitHub Releases in github/codeql-action
-    sinon.stub(actionsUtil, "isRunningLocalAction").resolves(true);
-    sinon.stub(api, "getApiClient").value(() => ({
-        repos: {
-            listReleases: sinon.stub().resolves(undefined),
-        },
-        paginate: sinon.stub().resolves([
-            {
-                assets: [
-                    {
-                        name: "cli-version-2.12.0.txt",
-                    },
-                ],
-                tag_name: "codeql-bundle-20230106",
-            },
-        ]),
-    }));
-    t.is(await codeql.findCodeQLBundleTagDotcomOnly("2.12.0", (0, logging_1.getRunnerLogger)(true)), "codeql-bundle-20230106");
-});
-(0, ava_1.default)("findCodeQLBundleTagDotcomOnly() errors if no GitHub Release matches marker file", async (t) => {
-    // Look for GitHub Releases in github/codeql-action
-    sinon.stub(actionsUtil, "isRunningLocalAction").resolves(true);
-    sinon.stub(api, "getApiClient").value(() => ({
-        repos: {
-            listReleases: sinon.stub().resolves(undefined),
-        },
-        paginate: sinon.stub().resolves([
-            {
-                assets: [
-                    {
-                        name: "cli-version-2.12.0.txt",
-                    },
-                ],
-                tag_name: "codeql-bundle-20230106",
-            },
-        ]),
-    }));
-    await t.throwsAsync(async () => await codeql.findCodeQLBundleTagDotcomOnly("2.12.1", (0, logging_1.getRunnerLogger)(true)), {
-        message: "Failed to find a CodeQL bundle release for CLI version 2.12.1.",
-    });
 });
 function stubToolRunnerConstructor() {
     const runnerObjectStub = sinon.createStubInstance(toolrunner.ToolRunner);
