+++ src/codeql.test.ts
@@ -1,5 +1,5 @@
 import * as fs from "fs";
-import * as path from "path";
+import path from "path";
 
 import * as toolrunner from "@actions/exec/lib/toolrunner";
 import * as toolcache from "@actions/tool-cache";
@@ -11,15 +11,14 @@ import nock from "nock";
 import * as sinon from "sinon";
 
 import * as actionsUtil from "./actions-util";
-import * as api from "./api-client";
 import { GitHubApiDetails } from "./api-client";
 import * as codeql from "./codeql";
 import { AugmentationProperties, Config } from "./config-utils";
-import * as defaults from "./defaults.json";
+import * as defaults from "./defaults.json"; // Referenced from codeql-action-sync-tool!
 import { Feature, featureConfig } from "./feature-flags";
 import { Language } from "./languages";
 import { getRunnerLogger } from "./logging";
-import { setupTests, setupActionsVars, createFeatures } from "./testing-utils";
+import { setupTests, createFeatures, setupActionsVars } from "./testing-utils";
 import * as util from "./util";
 import { initializeEnvironment } from "./util";
 
@@ -401,38 +400,6 @@ test("download codeql bundle from github ae endpoint", async (t) => {
   });
 });
 
-test("parse codeql bundle url version", (t) => {
-  t.deepEqual(
-    codeql.getCodeQLURLVersion(
-      "https://github.com/.../codeql-bundle-20200601/..."
-    ),
-    "20200601"
-  );
-});
-
-test("convert to semver", (t) => {
-  const tests = {
-    "20200601": "0.0.0-20200601",
-    "20200601.0": "0.0.0-20200601.0",
-    "20200601.0.0": "20200601.0.0",
-    "1.2.3": "1.2.3",
-    "1.2.3-alpha": "1.2.3-alpha",
-    "1.2.3-beta.1": "1.2.3-beta.1",
-  };
-
-  for (const [version, expectedVersion] of Object.entries(tests)) {
-    try {
-      const parsedVersion = codeql.convertToSemVer(
-        version,
-        getRunnerLogger(true)
-      );
-      t.deepEqual(parsedVersion, expectedVersion);
-    } catch (e) {
-      t.fail(e instanceof Error ? e.message : String(e));
-    }
-  }
-});
-
 test("getExtraOptions works for explicit paths", (t) => {
   t.deepEqual(codeql.getExtraOptions({}, ["foo"], []), []);
 
@@ -472,24 +439,6 @@ test("getExtraOptions throws for bad content", (t) => {
       []
     )
   );
-});
-
-test("getCodeQLActionRepository", (t) => {
-  const logger = getRunnerLogger(true);
-
-  initializeEnvironment("1.2.3");
-
-  // isRunningLocalAction() === true
-  delete process.env["GITHUB_ACTION_REPOSITORY"];
-  process.env["RUNNER_TEMP"] = path.dirname(__dirname);
-  const repoLocalRunner = codeql.getCodeQLActionRepository(logger);
-  t.deepEqual(repoLocalRunner, "github/codeql-action");
-
-  // isRunningLocalAction() === false
-  sinon.stub(actionsUtil, "isRunningLocalAction").returns(false);
-  process.env["GITHUB_ACTION_REPOSITORY"] = "xxx/yyy";
-  const repoEnv = codeql.getCodeQLActionRepository(logger);
-  t.deepEqual(repoEnv, "xxx/yyy");
 });
 
 test("databaseInterpretResults() does not set --sarif-add-query-help for 2.7.0", async (t) => {
@@ -927,60 +876,6 @@ test("databaseInterpretResults() does not set --sarif-add-baseline-file-info for
       "--sarif-add-baseline-file-info"
     ),
     "--sarif-add-baseline-file-info must be absent, but it is present"
-  );
-});
-
-test("findCodeQLBundleTagDotcomOnly() matches GitHub Release with marker file", async (t) => {
-  // Look for GitHub Releases in github/codeql-action
-  sinon.stub(actionsUtil, "isRunningLocalAction").resolves(true);
-  sinon.stub(api, "getApiClient").value(() => ({
-    repos: {
-      listReleases: sinon.stub().resolves(undefined),
-    },
-    paginate: sinon.stub().resolves([
-      {
-        assets: [
-          {
-            name: "cli-version-2.12.0.txt",
-          },
-        ],
-        tag_name: "codeql-bundle-20230106",
-      },
-    ]),
-  }));
-  t.is(
-    await codeql.findCodeQLBundleTagDotcomOnly("2.12.0", getRunnerLogger(true)),
-    "codeql-bundle-20230106"
-  );
-});
-
-test("findCodeQLBundleTagDotcomOnly() errors if no GitHub Release matches marker file", async (t) => {
-  // Look for GitHub Releases in github/codeql-action
-  sinon.stub(actionsUtil, "isRunningLocalAction").resolves(true);
-  sinon.stub(api, "getApiClient").value(() => ({
-    repos: {
-      listReleases: sinon.stub().resolves(undefined),
-    },
-    paginate: sinon.stub().resolves([
-      {
-        assets: [
-          {
-            name: "cli-version-2.12.0.txt",
-          },
-        ],
-        tag_name: "codeql-bundle-20230106",
-      },
-    ]),
-  }));
-  await t.throwsAsync(
-    async () =>
-      await codeql.findCodeQLBundleTagDotcomOnly(
-        "2.12.1",
-        getRunnerLogger(true)
-      ),
-    {
-      message: "Failed to find a CodeQL bundle release for CLI version 2.12.1.",
-    }
   );
 });
